add settings integrity check, change die/restart
To prevent extraneous and useless errors, the way that the program quits and/or restarts has been adjusted to rely on bat files. will add multiplatform support soon.

added a version setting to check if the settings have been updated, and prompt the user to reset to default as needed, this setting will only be updated when a change that would break settings occurs.