add warm/cool system
added warming and cooling system, currently requires user to be a mod, will rework permission system soon.

Added way to run the bot in test mode via command line arguments.

changed way that exiting is handled.

added temp_decay to settings with default of 900, this is value that will be used for the timer before the warm/cool roles are removed.