Implement role storage and restore
implemented on_member_remove to store the users roles if they are not currently frozen.

Does not store if user is currently frozen because when user is frozen, all roles are already stored for them. Thus including it here would be overly cumbersome.

When user rejoins server, the appropriate roles are re-added, or freeze is re-added.